+++ gson/src/test/java/feign/gson/GsonModuleTest.java
@@ -141,7 +141,7 @@ public class GsonModuleTest {
     DecoderBindings bindings = new DecoderBindings();
     ObjectGraph.create(bindings).inject(bindings);
 
-    Response response = Response.create(204, "OK", Collections.<String, Collection<String>>emptyMap(), null);
+    Response response = Response.create(204, "OK", Collections.<String, Collection<String>>emptyMap(), (byte[]) null);
     assertNull(bindings.decoder.decode(response, String.class));
   }