+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Version 7.1
 * Introduces feign.@Param to annotate template parameters. Users must migrate from `javax.inject.@Named` to `feign.@Param` before updating to Feign 8.0.
+* Adds OkHttp integration
 * Allows multiple headers with the same name.
 
 ### Version 7.0